♻️  refactor: file 이름 수정
old: AppMemberExternalApi -> new: MemberAppExternalApi
