+++ member-app-external-api/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'app-member-external-api'
+rootProject.name = 'member-app-external-api'
