+++ member-app-external-api/src/main/java/io/howstheairtoday/MemberAppExternalApiApplication.java
@@ -4,10 +4,10 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class AppMemberExternalApiApplication {
+public class MemberAppExternalApiApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(AppMemberExternalApiApplication.class, args);
+        SpringApplication.run(MemberAppExternalApiApplication.class, args);
     }
 
 }
