+++ member-app-external-api/src/main/java/io/howstheairtoday/memberappexternalapi/config/RootConfig.java
@@ -1,4 +1,4 @@
-package io.howstheairtoday.appmemberexternalapi.config;
+package io.howstheairtoday.memberappexternalapi.config;
 
 import org.modelmapper.ModelMapper;
 import org.modelmapper.convention.MatchingStrategies;
