+++ member-app-external-api/src/main/java/io/howstheairtoday/memberappexternalapi/service/handler/MemberLoginSuccessHandler.java
@@ -1,4 +1,4 @@
-package io.howstheairtoday.appmemberexternalapi.service.handler;
+package io.howstheairtoday.memberappexternalapi.service.handler;
 
 import java.io.IOException;
