+++ member-app-external-api/src/test/java/io/howstheairtoday/memberappexternalapi/MemberAppExternalApiApplicationTests.java
@@ -1,10 +1,10 @@
-package io.howstheairtoday.appmemberexternalapi;
+package io.howstheairtoday.memberappexternalapi;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class AppMemberExternalApiApplicationTests {
+class MemberAppExternalApiApplicationTests {
 
     @Test
     void contextLoads() {
