+++ module-core/build.gradle
@@ -29,6 +29,8 @@ dependencies {
     api 'org.springframework.data:spring-data-jpa'
     api 'org.modelmapper:modelmapper:3.1.0'
     api 'org.springframework.boot:spring-boot-starter-security'
+    api 'javax.xml.bind:jaxb-api:2.3.0'
+    api 'io.jsonwebtoken:jjwt:0.9.1'
     testImplementation 'org.springframework.security:spring-security-test'
 }
